requirements.txt
ultralytics==8.3.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
opencv-python>=4.8.0
pillow>=10.0.0
matplotlib>=3.7.0
scikit-image>=0.22.0
scipy>=1.11.0
